+++ .env.sample
@@ -1,2 +1,3 @@
 MYSQL_DB_URL=mysql://r00t:r00t@0.0.0.0:3308/unit_test
 POSTGRES_DB_URL=postgres://postgres:postgress@0.0.0.0:5438/unit_test
+MONGO_DB_URL=mongodb://root:root@0.0.0.0:27018
