+++ .npmignore
@@ -1,3 +1,4 @@
 .eslintrc.js
 .gitignore
 src/test
+build/test
