+++ package.json
@@ -1,13 +1,13 @@
 {
   "name": "unified-errors-handler",
-  "version": "3.1.3",
+  "version": "3.2.3",
   "description": "Unified Errors Handler is A Powerful Error Handling Library for Node.js that unify error structure across application. it can unify database errors.",
   "main": "build/index.js",
   "files": [
     "build/**/*"
   ],
   "scripts": {
-    "test": "mocha -r ts-node/register src/test/helpers/init.ts src/test/**/*.test.ts",
+    "test": "mocha -r ts-node/register src/test/helpers/init.ts src/test/**/**/*.test.ts",
     "build": "tsc",
     "format": "prettier --config .prettierrc 'src/**/*.ts' --write",
     "lint": "eslint . --ext .ts",
@@ -70,8 +70,10 @@
     "typescript": "^5.4.2"
   },
   "dependencies": {
+    "chalk": "^4.1.2",
     "db-errors": "^0.2.3",
     "http-status": "^1.7.4",
+    "moment": "^2.30.1",
     "ramda": "^0.30.1"
   }
 }
