+++ src/lib/exceptions/interfaces/exception.interface.ts
@@ -1,3 +1,5 @@
+import { ILoggerOptions } from '../../logger/logger.interface';
+
 export interface IException {
   message: string;
   code?: string;
@@ -5,15 +7,6 @@ export interface IException {
   details?: { [key: string]: any };
 }
 
-type ConsoleLoggerOptions = {
-  format?: string;
-};
-
-type LoggerTypes = { type: 'console'; options?: ConsoleLoggerOptions };
-
-export interface LoggingOptions {
-  logger?: LoggerTypes;
-}
 export interface IExceptionMapperOptions {
   mapDBExceptions?: boolean;
   parseSequelizeExceptions?: boolean;
@@ -21,5 +14,5 @@ export interface IExceptionMapperOptions {
   parseTypeORMExceptions?: boolean;
   parseObjectionJSExceptions?: boolean;
   parseKnexJSExceptions?: boolean;
-  logging?: LoggingOptions;
+  loggerOptions?: ILoggerOptions;
 }
