+++ src/test/helpers/modules/cutom-logger/custom-logger.ts
@@ -1,0 +1,9 @@
+import ILogger from '../../../../lib/logger/logger.interface';
+
+class CustomLogger implements ILogger {
+  log(error: any): void {
+    console.log(error.message);
+  }
+}
+
+export default new CustomLogger();
