+++ src/test/objection/mysql-objection.test.ts
@@ -1,20 +1,20 @@
 import assert from 'assert';
-import { PostgresDatabase } from './helpers/database/postgres-database';
-import { UserRepository } from './helpers/modules/postgres/objectionjs/users/user.repository';
-import exceptionMapper from '../lib/exceptions/exception-mapper';
-import { PetRepository } from './helpers/modules/postgres/objectionjs/pets/pet.repository';
+import exceptionMapper from '../../lib/exceptions/exception-mapper';
+import { MYSQLDatabase } from '../helpers/database/mysql-database';
+import { PetRepository } from '../helpers/modules/mysql/objectionjs/pets/pet.repository';
+import { UserRepository } from '../helpers/modules/mysql/objectionjs/users/user.repository';
 
-describe('Postgres ObjectionJS Testing', function () {
+describe('MYSQL ObjectionJS Testing', function () {
   let userRepo: UserRepository;
   let petRepo: PetRepository;
 
   before(async function () {
-    userRepo = new UserRepository(await PostgresDatabase.getInstance());
-    petRepo = new PetRepository(await PostgresDatabase.getInstance());
+    userRepo = new UserRepository(await MYSQLDatabase.getInstance());
+    petRepo = new PetRepository(await MYSQLDatabase.getInstance());
   });
 
   beforeEach(async function () {
-    await PostgresDatabase.seed();
+    await MYSQLDatabase.seed();
   });
 
   it('Should violate Unique Constraint.', async function () {
@@ -32,9 +32,12 @@ describe('Postgres ObjectionJS Testing', function () {
       const mappedError = exceptionMapper(e, {
         parseObjectionJSExceptions: true,
       }).serializeErrors();
-      assert.equal(mappedError?.[0]?.fields?.[0], 'name');
-      assert.equal(mappedError?.[0]?.code, 'DATA_ALREADY_EXIST');
-      assert.equal(mappedError?.[0]?.message, 'name already exist');
+      assert.deepEqual(mappedError, [
+        {
+          code: 'DATA_ALREADY_EXIST',
+          message: 'data already exist',
+        },
+      ]);
     }
   });
 
@@ -55,9 +58,8 @@ describe('Postgres ObjectionJS Testing', function () {
       }).serializeErrors();
       assert.deepEqual(mappedError, [
         {
-          fields: ['fname', 'lname'],
           code: 'DATA_ALREADY_EXIST',
-          message: 'fname, lname already exist',
+          message: 'data already exist',
         },
       ]);
     }
@@ -135,18 +137,17 @@ describe('Postgres ObjectionJS Testing', function () {
       ]);
     }
   });
-
   it('Should violate Check Constraint.', async function () {
     const userToBeInserted: any = {
       name: 'Osama',
       fname: 'Ahmed',
       lname: 'Adel',
       status: 'Active',
-      gender: 'FEMALEE',
-      age: 43,
+      gender: 'fff',
     };
     try {
       await userRepo.create(userToBeInserted);
+
       throw new Error('Database error must be fired.');
     } catch (e) {
       const mappedError = exceptionMapper(e, {
@@ -156,9 +157,6 @@ describe('Postgres ObjectionJS Testing', function () {
         {
           code: 'INVALID_VALUES',
           message: 'Invalid Values',
-          details: {
-            constraint: 'user_gender_check',
-          },
         },
       ]);
     }
