+++ src/test/sequelize/mysql-sequelize.test.ts
@@ -1,18 +1,10 @@
 import assert from 'assert';
-import exceptionMapper from '../lib/exceptions/exception-mapper';
-import { MYSQLDatabase } from './helpers/database/mysql-database';
-import { PetRepository } from './helpers/modules/mysql/objectionjs/pets/pet.repository';
-import { UserRepository } from './helpers/modules/mysql/objectionjs/users/user.repository';
+import exceptionMapper from '../../lib/exceptions/exception-mapper';
+import User from '../helpers/modules/mysql/sequelize/users/user.model';
+import Pet from '../helpers/modules/mysql/sequelize/pets/pet.model';
+import { MYSQLDatabase } from '../helpers/database/mysql-database';
 
-describe('MYSQL ObjectionJS Testing', function () {
-  let userRepo: UserRepository;
-  let petRepo: PetRepository;
-
-  before(async function () {
-    userRepo = new UserRepository(await MYSQLDatabase.getInstance());
-    petRepo = new PetRepository(await MYSQLDatabase.getInstance());
-  });
-
+describe('MYSQL Sequelize Testing', function () {
   beforeEach(async function () {
     await MYSQLDatabase.seed();
   });
@@ -26,22 +18,19 @@ describe('MYSQL ObjectionJS Testing', function () {
       age: 28,
     };
     try {
-      await userRepo.create(userToBeInserted);
+      await User.create(userToBeInserted);
       throw new Error('Database error must be fired.');
     } catch (e) {
       const mappedError = exceptionMapper(e, {
-        parseObjectionJSExceptions: true,
+        parseSequelizeExceptions: true,
       }).serializeErrors();
-      assert.deepEqual(mappedError, [
-        {
-          code: 'DATA_ALREADY_EXIST',
-          message: 'data already exist',
-        },
-      ]);
+      // assert.equal(mappedError?.[0]?.fields?.[0], 'name');
+      assert.equal(mappedError?.[0]?.code, 'DATA_ALREADY_EXIST');
+      assert.equal(mappedError?.[0]?.message, 'data already exist');
     }
   });
 
-  it('Should violate Unique Constraint (multi columns).', async function () {
+  it('Should violate Unique Constraint  (multi columns).', async function () {
     const userToBeInserted = {
       name: 'Osama',
       fname: 'Ahmed',
@@ -50,14 +39,15 @@ describe('MYSQL ObjectionJS Testing', function () {
       age: 28,
     };
     try {
-      await userRepo.create(userToBeInserted);
+      await User.create(userToBeInserted);
       throw new Error('Database error must be fired.');
     } catch (e) {
       const mappedError = exceptionMapper(e, {
-        parseObjectionJSExceptions: true,
+        parseSequelizeExceptions: true,
       }).serializeErrors();
       assert.deepEqual(mappedError, [
         {
+          // fields: ['fname', 'lname'],
           code: 'DATA_ALREADY_EXIST',
           message: 'data already exist',
         },
@@ -73,11 +63,11 @@ describe('MYSQL ObjectionJS Testing', function () {
       status: 'Active',
     };
     try {
-      await userRepo.create(user);
+      await User.create(user);
       throw new Error('Database error must be fired.');
     } catch (e) {
       const mappedError = exceptionMapper(e, {
-        parseObjectionJSExceptions: true,
+        parseSequelizeExceptions: true,
       }).serializeErrors();
       assert.deepEqual(mappedError, [
         {
@@ -98,11 +88,11 @@ describe('MYSQL ObjectionJS Testing', function () {
       type: 'Cow',
     };
     try {
-      await petRepo.create(pet);
+      await Pet.create(pet);
       throw new Error('Database error must be fired.');
     } catch (e) {
       const mappedError = exceptionMapper(e, {
-        parseObjectionJSExceptions: true,
+        parseSequelizeExceptions: true,
       }).serializeErrors();
       assert.deepEqual(mappedError, [
         {
@@ -119,11 +109,15 @@ describe('MYSQL ObjectionJS Testing', function () {
 
   it('Should violate Foreign Constraint (delete row has reference in another table).', async function () {
     try {
-      await userRepo.delete({ name: 'Ahmed' });
+      await User.destroy({
+        where: {
+          name: 'Ahmed',
+        },
+      });
       throw new Error('Database error must be fired.');
     } catch (e) {
       const mappedError = exceptionMapper(e, {
-        parseObjectionJSExceptions: true,
+        parseSequelizeExceptions: true,
       }).serializeErrors();
       assert.deepEqual(mappedError, [
         {
@@ -137,6 +131,7 @@ describe('MYSQL ObjectionJS Testing', function () {
       ]);
     }
   });
+
   it('Should violate Check Constraint.', async function () {
     const userToBeInserted: any = {
       name: 'Osama',
@@ -146,12 +141,11 @@ describe('MYSQL ObjectionJS Testing', function () {
       gender: 'fff',
     };
     try {
-      await userRepo.create(userToBeInserted);
-
+      await User.create(userToBeInserted);
       throw new Error('Database error must be fired.');
     } catch (e) {
       const mappedError = exceptionMapper(e, {
-        parseObjectionJSExceptions: true,
+        parseSequelizeExceptions: true,
       }).serializeErrors();
       assert.deepEqual(mappedError, [
         {
@@ -172,11 +166,11 @@ describe('MYSQL ObjectionJS Testing', function () {
       age: 999999999999999,
     };
     try {
-      await userRepo.create(userToBeInserted);
+      await User.create(userToBeInserted);
       throw new Error('Database error must be fired.');
     } catch (e) {
       const mappedError = exceptionMapper(e, {
-        parseObjectionJSExceptions: true,
+        parseSequelizeExceptions: true,
       }).serializeErrors();
       assert.deepEqual(mappedError, [
         {
