+++ src/test/sequelize/pg-sequelize.test.ts
@@ -1,36 +1,28 @@
 import assert from 'assert';
-import { PostgresDatabase } from './helpers/database/postgres-database';
-import exceptionMapper from '../lib/exceptions/exception-mapper';
-import User from './helpers/modules/postgres/typeorm/users/user.model';
-import { Repository } from 'typeorm';
-import Pet from './helpers/modules/postgres/typeorm/pets/pet.model';
+import { PostgresDatabase } from '../helpers/database/postgres-database';
+import exceptionMapper from '../../lib/exceptions/exception-mapper';
+import User from '../helpers/modules/postgres/sequelize/users/user.model';
+import Pet from '../helpers/modules/postgres/sequelize/pets/pet.model';
 
-describe('Postgres Typeorm Testing', function () {
-  let userRepo: Repository<User>;
-  let petRepo: Repository<Pet>;
-
-  before(async function () {
-    userRepo = (await PostgresDatabase.getTypeormInstance()).getRepository(User);
-    petRepo = (await PostgresDatabase.getTypeormInstance()).getRepository(Pet);
-  });
-
+describe('Postgres Sequelize Testing', function () {
   beforeEach(async function () {
     await PostgresDatabase.seed();
   });
 
   it('Should violate Unique Constraint.', async function () {
-    const user = new User();
-    user.name = 'Ahmed';
-    user.fname = 'Hassan';
-    user.lname = 'ALi';
-    user.status = 'Active';
-    user.age = 28;
+    const userToBeInserted = {
+      name: 'Ahmed',
+      fname: 'Hassan',
+      lname: 'ALi',
+      status: 'Active',
+      age: 28,
+    };
     try {
-      await userRepo.save(user);
+      await User.create(userToBeInserted);
       throw new Error('Database error must be fired.');
     } catch (e) {
       const mappedError = exceptionMapper(e, {
-        parseTypeORMExceptions: true,
+        parseSequelizeExceptions: true,
       }).serializeErrors();
       assert.equal(mappedError?.[0]?.fields?.[0], 'name');
       assert.equal(mappedError?.[0]?.code, 'DATA_ALREADY_EXIST');
@@ -38,7 +30,7 @@ describe('Postgres Typeorm Testing', function () {
     }
   });
 
-  it('Should violate Unique Constraint (multi columns).', async function () {
+  it('Should violate Unique Constraint  (multi columns).', async function () {
     const userToBeInserted = {
       name: 'Osama',
       fname: 'Ahmed',
@@ -47,11 +39,11 @@ describe('Postgres Typeorm Testing', function () {
       age: 28,
     };
     try {
-      await userRepo.save(userToBeInserted);
+      await User.create(userToBeInserted);
       throw new Error('Database error must be fired.');
     } catch (e) {
       const mappedError = exceptionMapper(e, {
-        parseTypeORMExceptions: true,
+        parseSequelizeExceptions: true,
       }).serializeErrors();
       assert.deepEqual(mappedError, [
         {
@@ -71,11 +63,11 @@ describe('Postgres Typeorm Testing', function () {
       status: 'Active',
     };
     try {
-      await userRepo.save(user);
+      await User.create(user);
       throw new Error('Database error must be fired.');
     } catch (e) {
       const mappedError = exceptionMapper(e, {
-        parseTypeORMExceptions: true,
+        parseSequelizeExceptions: true,
       }).serializeErrors();
       assert.deepEqual(mappedError, [
         {
@@ -96,11 +88,11 @@ describe('Postgres Typeorm Testing', function () {
       type: 'Cow',
     };
     try {
-      await petRepo.save(pet);
+      await Pet.create(pet);
       throw new Error('Database error must be fired.');
     } catch (e) {
       const mappedError = exceptionMapper(e, {
-        parseTypeORMExceptions: true,
+        parseSequelizeExceptions: true,
       }).serializeErrors();
       assert.deepEqual(mappedError, [
         {
@@ -117,13 +109,15 @@ describe('Postgres Typeorm Testing', function () {
 
   it('Should violate Foreign Constraint (delete row has reference in another table).', async function () {
     try {
-      await userRepo.delete({
-        name: 'Ahmed',
+      await User.destroy({
+        where: {
+          name: 'Ahmed',
+        },
       });
       throw new Error('Database error must be fired.');
     } catch (e) {
       const mappedError = exceptionMapper(e, {
-        parseTypeORMExceptions: true,
+        parseSequelizeExceptions: true,
       }).serializeErrors();
       assert.deepEqual(mappedError, [
         {
@@ -148,11 +142,11 @@ describe('Postgres Typeorm Testing', function () {
       age: 23,
     };
     try {
-      await userRepo.save(userToBeInserted);
+      await User.create(userToBeInserted);
       throw new Error('Database error must be fired.');
     } catch (e) {
       const mappedError = exceptionMapper(e, {
-        parseTypeORMExceptions: true,
+        parseSequelizeExceptions: true,
       }).serializeErrors();
       assert.deepEqual(mappedError, [
         {
@@ -173,14 +167,14 @@ describe('Postgres Typeorm Testing', function () {
       lname: 'Adel',
       status: 'Active',
       gender: 'FEMALE',
-      age: '9999999999999999999999999999',
+      age: 999999999999999,
     };
     try {
-      await userRepo.save(userToBeInserted);
+      await User.create(userToBeInserted);
       throw new Error('Database error must be fired.');
     } catch (e) {
       const mappedError = exceptionMapper(e, {
-        parseTypeORMExceptions: true,
+        parseSequelizeExceptions: true,
       }).serializeErrors();
       assert.deepEqual(mappedError, [
         {
